opslaan

+++ app/src/main/java/main/WTLibraryApp/Book/Copy/CopyRepository.java
@@ -16,10 +16,12 @@ public interface CopyRepository extends JpaRepository<Copy, CopyPK>{
 	
 	// findById method for combined-id object
 	List<Copy> findByCopyPkId(CopyPK copyPk);
+	
 	// deleteById method for combined-id object
 //	Copy deleteByCopyPkId(CopyPK copyPk);
+	
 	// findById method for user-id object
-	List<Copy> findCopyByUserId(int id);
+	List<Copy> findCopyByUserId(long id);
 	
 	//search copies by bookId
 	@Query(value = "select * from copies where book_id like %:id%", nativeQuery = true)
